+++ noodles-bgzf/src/io.rs
@@ -1,0 +1,5 @@
+//! BGZF I/O.
+
+mod read;
+
+pub use self::read::Read;