+++ noodles-bgzf/src/io.rs
@@ -1,5 +1,6 @@
 //! BGZF I/O.
 
+mod buf_read;
 mod read;
 
-pub use self::read::Read;
+pub use self::{buf_read::BufRead, read::Read};