+++ noodles-bgzf/src/io/buf_read.rs
@@ -1,0 +1,4 @@
+use std::io;
+
+/// A buffered BGZF reader.
+pub trait BufRead: io::BufRead + super::Read {}