+++ noodles-bgzf/src/io.rs
@@ -2,5 +2,6 @@
 
 mod buf_read;
 mod read;
+mod seek;
 
-pub use self::{buf_read::BufRead, read::Read};
+pub use self::{buf_read::BufRead, read::Read, seek::Seek};