+++ noodles-bgzf/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.29.0 - 2024-05-02
 
 ### Added